timber_locator/requirements.txt
Django>=5.2.1
django-crispy-forms>=2.1
crispy-bootstrap5>=2023.10
django-mptt>=0.15.0
django-htmx>=1.17.3 